legalEze-backend/legaleze/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-caching>=2.1.0
flask-limiter>=3.5.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
pymupdf>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
flasgger>=0.9.7.1
redis>=5.0.0
uuid>=1.30
flask-migrate>=4.0.0
